put all content from header to sidebar and page title [skip ci]
