+++ frontend/src/component/TabBar/TabBar.jsx
@@ -22,6 +22,7 @@ export const TabBar = props => {
         icon={spaceType.faIcon}
         breadcrumbsList={props.breadcrumbs}
         iconTooltip={spaceType.label}
+        // emailNotifActivated={props.system.config.email_notification_activated}
       />
 
       <div className='tabBar__tabs'>
