+++ frontend/src/container/AdvancedSearch.jsx
@@ -463,6 +463,7 @@ export class AdvancedSearch extends React.Component {
               )}
               icon='fas fa-search'
               breadcrumbsList={props.breadcrumbs}
+              // emailNotifActivated={props.system.config.email_notification_activated}
             />
 
             <PageContent parentClass={classnames('advancedSearch', { advancedSearch__openMenu: state.isFilterMenuOpen })}>
