+++ frontend/src/container/JoinWorkspace.jsx
@@ -183,6 +183,7 @@ export class JoinWorkspace extends React.Component {
               title={props.t('Join a space')}
               icon='fas fa-users'
               breadcrumbsList={props.breadcrumbs}
+              // emailNotifActivated={props.system.config.email_notification_activated}
             />
 
             <PageContent parentClass={`${className}__content`}>
