+++ frontend/src/container/SimpleSearch.jsx
@@ -207,6 +207,7 @@ export class SimpleSearch extends React.Component {
               )}
               icon='fas fa-search'
               breadcrumbsList={props.breadcrumbs}
+              // emailNotifActivated={props.system.config.email_notification_activated}
             />
             {state.isLoading
               ? <Loading />
