+++ frontend_app_admin_workspace_user/src/component/AdminForm.jsx
@@ -112,7 +112,7 @@ export class AdminForm extends React.Component {
           parentClass='adminUser'
           title='Formulaires manager'
           icon='users'
-
+          // emailNotifActivated={props.system.config.email_notification_activated}
         />
 
         <PageContent parentClass='adminUser'>
