+++ frontend_app_admin_workspace_user/src/component/AdminUser.jsx
@@ -114,6 +114,7 @@ export class AdminUser extends React.Component {
           title={props.t('User account management')}
           icon='fas fa-users'
           breadcrumbsList={props.breadcrumbsList}
+          // emailNotifActivated={props.system.config.email_notification_activated}
         />
 
         <PageContent parentClass='adminUser'>
